Organized some examples.
